requirements.txt
flask
python-dotenv
langchain
langchain_community
langgraph
langchain-openai
langchain-google-genai
faiss-cpu
pandas
sqlalchemy
pydantic
requests
